Create working logout page

+++ app/logout/page.tsx
@@ -1,34 +1,55 @@
 'use client';
-import { ReactNode, useEffect, useContext } from 'react';
+import { ReactNode, useEffect, useContext, useState } from 'react';
 import { Spinner } from '@heroui/react';
 import toast from 'react-hot-toast';
 import { useRouter } from 'next/navigation';
 import { createInitStore, StoreContext } from '@/utils/store';
 import { useLocalStorageSync } from '@/utils/hooks/useLocalStorageSync';
+import { makeBackEndRequest, BASE_URL } from '@/utils';
+import { HTTP_METHODS } from '@/utils/ts';
 
-export default function HomePage(): ReactNode {
+export default function LogoutPage(): ReactNode {
   useLocalStorageSync();
   const { store, updateStore } = useContext(StoreContext);
   const router = useRouter();
 
-  console.log('store in logout ', store);
-  console.log('\n');
+  const [hasLoggedOut, setHasLoggedOut] = useState(false);
 
   useEffect(() => {
-    if (typeof window !== 'undefined') {
-      localStorage.clear();
+    if (hasLoggedOut || !store.user || !updateStore) return;
 
-      if (updateStore) {
+    const logoutOfApp = async () => {
+      const logoutURL = `${BASE_URL}/auth/logout`;
+
+      if (typeof window !== 'undefined' && store.user) {
+        localStorage.clear();
+
+        const { firstName, lastName, email } = store.user;
+        const name = `${firstName} ${lastName}`;
         const storeReset = createInitStore();
-        updateStore((prev) => ({ ...prev, ...storeReset }));
+
+        updateStore(storeReset);
+
+        const reqPayload = {
+          apiURL: logoutURL,
+          method: 'POST' as HTTP_METHODS,
+          bodyPayload: { name, email }
+        };
+
+        await makeBackEndRequest<Response>(reqPayload, false);
 
         toast.success("You've been logged out. See you later. 👋🏼", {
           duration: 3000
         });
-        router.push('/');
+
+        setHasLoggedOut(true);
+
+        setTimeout(() => router.push('/'), 3000);
       }
-    }
-  }, [router, updateStore]);
+    };
+
+    logoutOfApp();
+  }, [router, store.user, updateStore, hasLoggedOut]);
 
   return (
     <div className="w-[100%] min-h-screen flex justify-center items-center">
